pyproject.toml
[project]
name = "agent-orch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0.2",
    "google-adk>=1.18.0",
    "google-genai>=1.49.0",
    "httpx>=0.28.1",
    "litellm>=1.79.1",
    "mcp>=1.21.0",
    "python-dotenv>=1.2.1",
]
